fifth commit

+++ src/main/java/com/task/service/dto/EmployeeDto.java
@@ -6,7 +6,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.time.LocalDate;
 import java.util.List;
 
-public record EmployeeDto(Integer id, String code,
+public record EmployeeDto(Long id, String code,
                           String name,
                           LocalDate dateOfBirth,
                           AddressDto addressDto,
